refactor: 도메인 객체와 테스트의 패키지 경로 같도록 변경
