+++ src/test/java/racingCar/domains/NameTest.java
@@ -1,8 +1,7 @@
-package racingCar;
+package racingCar.domains;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingCar.domains.Name;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
