+++ src/test/java/racingCar/domains/RacingTest.java
@@ -1,8 +1,7 @@
-package racingCar;
+package racingCar.domains;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingCar.domains.Racing;
 
 import static org.assertj.core.api.Assertions.assertThat;
